requirements.txt
langchain
langchain-core
langchain-community
langchain-google-genai

python-dotenv
streamlit
unstructured
tiktoken
faiss-cpu
libmagic
python-magic
python-magic-bin

google-api-core
tenacity
requests
streamlit-option-menu
pathlib
